requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0

# Google APIs
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
Faker==30.3.0

